+++ src/cards/card_deck.rs
@@ -260,6 +260,8 @@ where
 
     /// Returns the probability of the cards to be drawn.
     ///
+    /// The probabilites are guaranteed to be `> 0`.
+    /// 
     /// # Example
     ///
     /// ```
@@ -281,7 +283,8 @@ where
         let size = self.size();
         self.cards
             .iter()
-            .map(|(card, count)| (card, Probability::new(*count, size)))
+            .filter(|(_, &count)| count > 0)
+            .map(|(card, &count)| (card, Probability::new(count, size)))
             .collect()
     }
 
@@ -348,5 +351,18 @@ where
         let mut deck = (*self).clone();
         deck.remove_times(card, 1);
         deck
+    }
+}
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn probabilities_no_zero_returns() {
+        let mut deck = CardDeck::from(vec![1,2,3]);
+        deck.remove_times(3, 1);
+        assert!(deck.probabilities().values().all(|&x| x > PROBABILITY_ZERO));
     }
 }